+++ sgoop/utilities.py
